[pnnl/SHAD#63] small refactoring of shad hash
