+++ include/shad/core/type_traits.h
@@ -1,0 +1,43 @@
+//===------------------------------------------------------------*- C++ -*-===//
+//
+//                                     SHAD
+//
+//      The Scalable High-performance Algorithms and Data Structure Library
+//
+//===----------------------------------------------------------------------===//
+//
+// Copyright 2018 Battelle Memorial Institute
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy
+// of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef INCLUDE_SHAD_CORE_TYPE_TRAITS_H
+#define INCLUDE_SHAD_CORE_TYPE_TRAITS_H
+
+#include <functional>
+#include <type_traits>
+
+namespace shad {
+
+template <class T>
+struct is_std_hashable :
+         std::integral_constant<bool,
+                                std::is_arithmetic<T>::value ||
+                                std::is_pointer<T>::value ||
+                                std::is_same<T, std::string>::value>{
+};
+
+}  // namespace shad
+
+#endif /* INCLUDE_SHAD_CORE_TYPE_TRAITS_H */
