+++ include/shad/data_structures/compare_and_hash_utils.h
@@ -32,6 +32,8 @@
 #include <vector>
 #include <type_traits>
 
+#include "shad/core/type_traits.h"
+
 namespace shad {
 
 /// @brief Comparison functor.
@@ -172,14 +174,6 @@ uint64_t HashFunction(const std::vector<KeyTy> &key, uint8_t seed) {
   hash += (hash << 15);
   return hash;
 }
-
-template <class T>
-struct is_std_hashable :
-         std::integral_constant<bool,
-                                std::is_arithmetic<T>::value ||
-                                std::is_pointer<T>::value ||
-                                std::is_same<T, std::string>::value>{
-};
 
 template <typename Key, bool=is_std_hashable<Key>::value>
 struct hash {
